requirements.txt
# Core Dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# AI/ML Models
openai-whisper==20231117
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Language Processing
googletrans==4.0.0rc1
langdetect==1.0.9
spacy==3.7.2

# Data Processing
scikit-learn==1.3.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
plotly==5.17.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 
